% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howfar-package.R
\docType{package}
\name{howfar-package}
\alias{howfar}
\alias{howfar-package}
\title{howfar: Fast Euclidean Distance Rasters}
\description{
Computes exact Euclidean distance rasters from vector spatial features
(points, lines, polygons) using a parallel linear-time distance transform
implemented in Rust.
}
\author{
\strong{Maintainer}: Roozbeh Valavi \email{valavi.r@gmail.com} (\href{https://orcid.org/0000-0003-2495-5277}{ORCID})

}
\keyword{internal}
